requirements.txt
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
mysqlclient==2.2.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
